format strings

+++ common/phone/obex/obexgen.c
@@ -1163,7 +1163,7 @@ GSM_Error OBEXGEN_GetTextFile(GSM_StateMachine *s, const char *FileName, char **
 /**
  * Sets single file on filesystem, file can be created or updated.
  */
-GSM_Error OBEXGEN_SetFile(GSM_StateMachine *s, const char *FileName, unsigned char *Buffer, int Length)
+GSM_Error OBEXGEN_SetFile(GSM_StateMachine *s, const char *FileName, unsigned char *Buffer, size_t Length)
 {
 	GSM_Error	error = ERR_NONE;
 	GSM_File 	File;
@@ -1739,7 +1739,7 @@ GSM_Error OBEXGEN_AddMemory(GSM_StateMachine *s, GSM_MemoryEntry *Entry)
 		error = OBEXGEN_InitPbLUID(s);
 		if (error != ERR_NONE) return error;
 
-		smprintf(s,"Adding phonebook entry %d:\n%s\n", size, req);
+		smprintf(s,"Adding phonebook entry %zd:\n%s\n", size, req);
 		Priv->UpdatePbLUID = true;
 		error = OBEXGEN_SetFile(s, "telecom/pb/luid/.vcf", req, size);
 		Entry->Location = Priv->PbLUIDCount;
@@ -1750,7 +1750,7 @@ GSM_Error OBEXGEN_AddMemory(GSM_StateMachine *s, GSM_MemoryEntry *Entry)
 		if (error != ERR_NONE) return error;
 
 		Entry->Location = OBEXGEN_GetFirstFreeLocation(&Priv->PbIndex, &Priv->PbIndexCount);
-		smprintf(s,"Adding phonebook entry %d at location %d:\n%s\n", size, Entry->Location, req);
+		smprintf(s,"Adding phonebook entry %zd at location %d:\n%s\n", size, Entry->Location, req);
 		sprintf(path, "telecom/pb/%d.vcf", Entry->Location);
 		error = OBEXGEN_SetFile(s, path, req, size);
 		return error;
@@ -2168,7 +2168,7 @@ GSM_Error OBEXGEN_AddCalendar(GSM_StateMachine *s, GSM_CalendarEntry *Entry)
 		error = OBEXGEN_InitCalLUID(s);
 		if (error != ERR_NONE) return error;
 
-		smprintf(s,"Adding calendar entry %d:\n%s\n", size, req);
+		smprintf(s,"Adding calendar entry %zd:\n%s\n", size, req);
 		Priv->UpdateCalLUID = true;
 		error = OBEXGEN_SetFile(s, "telecom/cal/luid/.vcs", req, size);
 		Entry->Location = Priv->CalLUIDCount;
@@ -2179,7 +2179,7 @@ GSM_Error OBEXGEN_AddCalendar(GSM_StateMachine *s, GSM_CalendarEntry *Entry)
 		if (error != ERR_NONE) return error;
 
 		Entry->Location = OBEXGEN_GetFirstFreeLocation(&Priv->CalIndex, &Priv->CalIndexCount);
-		smprintf(s,"Adding calendar entry %d at location %d:\n%s\n", size, Entry->Location, req);
+		smprintf(s,"Adding calendar entry %zd at location %d:\n%s\n", size, Entry->Location, req);
 		sprintf(path, "telecom/cal/%d.vcf", Entry->Location);
 		error = OBEXGEN_SetFile(s, path, req, size);
 		return error;
@@ -2545,7 +2545,7 @@ GSM_Error OBEXGEN_AddTodo(GSM_StateMachine *s, GSM_ToDoEntry *Entry)
 		error = OBEXGEN_InitCalLUID(s);
 		if (error != ERR_NONE) return error;
 
-		smprintf(s,"Adding todo entry %d:\n%s\n", size, req);
+		smprintf(s,"Adding todo entry %zd:\n%s\n", size, req);
 		Priv->UpdateTodoLUID = true;
 		error = OBEXGEN_SetFile(s, "telecom/cal/luid/.vcs", req, size);
 		Entry->Location = Priv->TodoLUIDCount;
@@ -2556,7 +2556,7 @@ GSM_Error OBEXGEN_AddTodo(GSM_StateMachine *s, GSM_ToDoEntry *Entry)
 		if (error != ERR_NONE) return error;
 
 		Entry->Location = OBEXGEN_GetFirstFreeLocation(&Priv->TodoIndex, &Priv->TodoIndexCount);
-		smprintf(s,"Adding todo entry %d at location %d:\n%s\n", size, Entry->Location, req);
+		smprintf(s,"Adding todo entry %zd at location %d:\n%s\n", size, Entry->Location, req);
 		sprintf(path, "telecom/cal/%d.vcf", Entry->Location);
 		error = OBEXGEN_SetFile(s, path, req, size);
 		return error;
@@ -2941,7 +2941,7 @@ GSM_Error OBEXGEN_AddNote(GSM_StateMachine *s, GSM_NoteEntry *Entry)
 		error = OBEXGEN_InitNoteLUID(s);
 		if (error != ERR_NONE) return error;
 
-		smprintf(s,"Adding note entry %d:\n%s\n", size, req);
+		smprintf(s,"Adding note entry %zd:\n%s\n", size, req);
 		Priv->UpdateNoteLUID = true;
 		error = OBEXGEN_SetFile(s, "telecom/nt/luid/.vnt", req, size);
 		Entry->Location = Priv->NoteLUIDCount;
@@ -2952,7 +2952,7 @@ GSM_Error OBEXGEN_AddNote(GSM_StateMachine *s, GSM_NoteEntry *Entry)
 		if (error != ERR_NONE) return error;
 
 		Entry->Location = OBEXGEN_GetFirstFreeLocation(&Priv->NoteIndex, &Priv->NoteIndexCount);
-		smprintf(s,"Adding note entry %d at location %d:\n%s\n", size, Entry->Location, req);
+		smprintf(s,"Adding note entry %zd at location %d:\n%s\n", size, Entry->Location, req);
 		sprintf(path, "telecom/nt/%d.vcf", Entry->Location);
 		error = OBEXGEN_SetFile(s, path, req, size);
 		return error;
